requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
pandas==2.2.3
openpyxl==3.1.5                     # for sending excel
